# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Çakışma"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcı oluştur (Admin)",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponseDto"
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    },
                    "totalCount": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize",
                    "totalCount",
                    "totalPages"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcıları listele (paged)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcıyı getir",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "id uyuşmazlığı"
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcı güncelle",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcı sil",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "TransferPulse API",
    "description": "TransferPulse backend (.NET → NestJS migration)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "profilePic": {
            "type": "string"
          },
          "isMailConfirm": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Banned",
              "Suspended"
            ]
          },
          "favouriteTeam": {
            "type": "string"
          },
          "reputationScore": {
            "type": "number"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "nickname",
          "isMailConfirm",
          "status",
          "reputationScore",
          "role",
          "createdAt"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "alice",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "example": "alice@acme.com",
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "example": "Secret123",
            "minLength": 8,
            "maxLength": 72
          },
          "nickname": {
            "type": "string",
            "example": "Alice",
            "maxLength": 50
          },
          "profilePic": {
            "type": "string",
            "maxLength": 500
          },
          "favouriteTeam": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "nickname"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Route :id ile eşleşmeli (uyuşmazsa 400)"
          },
          "nickname": {
            "type": "string",
            "maxLength": 50
          },
          "profilePic": {
            "type": "string",
            "maxLength": 500
          },
          "favouriteTeam": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}